+++ src/Attributes/Permission.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Iankibet\Streamline\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS)]
+class Permission
+{
+
+}
