--- src/Permissions/StreamlinePermission.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Iankibet\Streamline\Permissions;
-
-class StreamlinePermission
-{
-
-}
